MultiTenantDao Adapter

+++ contracts/dar/IWeightedRegistry.sol
@@ -5,7 +5,7 @@ interface IWeightedRegistry {
   function totalWeight() public view returns (uint256);
   function isWeighted() public view returns (bool);
 
-  function weightOfAsset(uint256 assetId) public view returns (uint64);
+  function weightOfAsset(uint256 assetId) public view returns (uint256);
   function weightOfHolder(address holder) public view returns (uint256);
 
   /**
@@ -15,11 +15,11 @@ interface IWeightedRegistry {
     address indexed from,
     address indexed to,
     uint256 indexed assetId,
-    uint64 weight
+    uint256 weight
   );
   event ChangeWeight(
     uint256 indexed assetId,
-    uint64 weight
+    uint256 weight
   );
 
 }
