+++ contracts/dar/WeightedAssetRegistryStorage.sol
@@ -12,7 +12,7 @@ contract WeightedAssetRegistryStorage is AssetRegistryStorage {
   /**
    * Stores the weight owned by a given asset
    */
-  mapping(uint256 => uint64) internal _weightOfAsset;
+  mapping(uint256 => uint256) internal _weightOfAsset;
 
   /**
    * Stores the total weight owned by a given account
